+++ src/main/java/com/example/api/repository/AdRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.api.model.Ad;
 
-public interface AdRepository extends JpaRepository<Ad, Integer> {
+public interface AdRepository extends JpaRepository<Ad, Long> {
     
 }
